+++ src/cryptography/x509/__init__.py
@@ -7,19 +7,19 @@ from __future__ import absolute_import, division, print_function
 from cryptography.x509.base import (
     Certificate, CertificateBuilder, CertificateRevocationList,
     CertificateSigningRequest, CertificateSigningRequestBuilder,
-    Extension, ExtensionType, GeneralNames,
-    InvalidVersion, IssuerAlternativeName,
-    ObjectIdentifier, RevokedCertificate, SubjectAlternativeName,
-    Version, load_der_x509_certificate,
-    load_der_x509_csr, load_pem_x509_certificate, load_pem_x509_csr,
+    InvalidVersion, RevokedCertificate,
+    Version, load_der_x509_certificate, load_der_x509_csr,
+    load_pem_x509_certificate, load_pem_x509_csr,
 )
 from cryptography.x509.extensions import (
     AccessDescription, AuthorityInformationAccess,
     AuthorityKeyIdentifier, BasicConstraints, CRLDistributionPoints,
     CertificatePolicies, DistributionPoint, DuplicateExtension,
-    ExtendedKeyUsage, ExtensionNotFound, Extensions, InhibitAnyPolicy,
-    KeyUsage, NameConstraints, NoticeReference, OCSPNoCheck, PolicyInformation,
-    ReasonFlags, SubjectKeyIdentifier, UnsupportedExtension, UserNotice
+    ExtendedKeyUsage, Extension, ExtensionNotFound, ExtensionType, Extensions,
+    GeneralNames, InhibitAnyPolicy, IssuerAlternativeName, KeyUsage,
+    NameConstraints, NoticeReference, OCSPNoCheck, PolicyInformation,
+    ReasonFlags, SubjectAlternativeName, SubjectKeyIdentifier,
+    UnsupportedExtension, UserNotice
 )
 from cryptography.x509.general_name import (
     DNSName, DirectoryName, GeneralName, IPAddress, OtherName, RFC822Name,
@@ -29,8 +29,8 @@ from cryptography.x509.general_name import (
 from cryptography.x509.name import Name, NameAttribute
 from cryptography.x509.oid import (
     AuthorityInformationAccessOID, CRLExtensionOID, CertificatePoliciesOID,
-    ExtendedKeyUsageOID, ExtensionOID, NameOID, SignatureAlgorithmOID,
-    _SIG_OIDS_TO_HASH
+    ExtendedKeyUsageOID, ExtensionOID, NameOID, ObjectIdentifier,
+    SignatureAlgorithmOID, _SIG_OIDS_TO_HASH
 )
